requirements.txt
djangorestframework
django-heroku>=0.3.1
dj-database-url>=0.5.0
Django>=3.0.2
gunicorn>=20.0.4
psycopg2>=2.8.3
pytz>=2019.1
whitenoise>=5.0